requirements.txt
aiofiles==0.7.0
click==7.1.2
fastapi==0.65.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
psycopg2-binary==2.8.6
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
requests==2.25.1
pytest==6.2.4